+++ test/test_rewrite.rb
@@ -41,4 +41,18 @@ describe Jellyfish::Rewrite do
     expect(call(app, '/top/from/inner')).eq '/top!/to/inner'
     expect(call(app, '/top/from/outer')).eq '/top!/to/outer'
   end
+
+  would 'map to with host and handle SCRIPT_NAME properly' do
+    app = Jellyfish::Builder.app do
+      map '/path', to: '/path', host: 'host' do
+        run lambda{ |env|
+          [200, {},
+            ["#{env['HTTP_HOST']} #{env['SCRIPT_NAME']} #{env['PATH_INFO']}"]]
+        }
+      end
+    end
+
+    expect(get('/path', app, 'HTTP_HOST' => 'host').dig(-1, -1)).
+      eq 'host  /path'
+  end
 end